// src/services/generated/tainacanV2.schemas.ts
export type FilterMetadatumMetadataTypeObject = {
  primitive_type?: string;
  options?: FilterMetadatumMetadataTypeObjectOptions;
};

/**
 * Nested metadatum summary on a filter (Filter._toArray).
 */
export interface FilterMetadatum {
  metadatum_id?: number | string;
  metadatum_name?: string;
  metadata_type_object?: FilterMetadatumMetadataTypeObject;
}

export type FilterEnabled = typeof FilterEnabled[keyof typeof FilterEnabled];


export const FilterEnabled = {
  yes: 'yes',
  no: 'no',
} as const;

/**
 * The filter type options
 */
export type FilterFilterTypeOptions = unknown[] | { [key: string]: unknown } | string;

export type FilterFilterTypeObject = { [key: string]: unknown };

export interface Filter {
  /** Unique identifier */
  id?: number;
  /** Status for control of visibility and access. */
  status?: string;
  /** Name of the filter */
  name?: string;
  /** Filter order. This metadata is used if filters were manually ordered. */
  order?: string | number;
  /** The filter description */
  description?: string;
  /** The filter type options */
  filter_type_options?: FilterFilterTypeOptions;
  /** The filter type class name, such as filter_type: Tainacan\Filter_Types\Checkbox */
  filter_type?: string;
  /** With this option enabled, the filter will appear as a button with an add icon, that should be pressed prior to loading any facet information. */
  begin_with_filter_collapsed?: string;
  /** With this option enabled, the filter will appear even in repository level items lists, such as the complete items list and the term items list. */
  display_in_repository_level_lists?: string;
  collection_id?: number | string;
  metadatum_id?: number | string;
  /** The maximum number of options to be loaded by default on the filter. */
  max_options?: number | string;
  /** The filter input placeholder. This is a simple message that will appear inside textual input and may indicate to the user what kind of information is expected. */
  placeholder?: string;
  /** Whether the filter description should be displayed below the input label instead of inside a tooltip. */
  description_bellow_name?: string;
  /** Whether current user can edit this object */
  current_user_can_edit?: boolean;
  /** Whether current user can delete this object */
  current_user_can_delete?: boolean;
  metadatum?: FilterMetadatum;
  filter_type_object?: FilterFilterTypeObject;
  inherited?: boolean;
  enabled?: FilterEnabled;
}

export type Filters = Filter[];

export type TermThumbnail = { [key: string]: unknown } | string | boolean;

export type TermHeaderImage = { [key: string]: unknown } | string | boolean;

export interface Term {
  /** Unique identifier for the object. */
  id?: number;
  /** Status for control of visibility and access. */
  status?: string;
  /** Unique identifier */
  term_id?: number;
  /** Name of the term */
  name?: string;
  /** @nullable */
  parent?: number | string | null;
  /** The term description */
  description?: string;
  /** The term taxonomy */
  taxonomy?: string;
  user?: number | string;
  /** @nullable */
  header_image_id?: number | string | null;
  /** @nullable */
  cover_page_id?: number | string | null;
  /** @nullable */
  hide_empty?: boolean | null;
  /** Whether current user can edit this object */
  current_user_can_edit?: boolean;
  /** Whether current user can delete this object */
  current_user_can_delete?: boolean;
  url?: string;
  thumbnail?: TermThumbnail;
  thumbnail_alt?: string | boolean;
  thumbnail_id?: number | string;
  header_image?: TermHeaderImage;
  hierarchy_path?: string;
  total_children?: number | string;
}